disable reports entirely.
