+++ src/eressea.c
@@ -28,11 +28,10 @@
 
 #include "calendar.h"
 #include "chaos.h"
-#include "creport.h"
 #include "items.h"
-#include "jsreport.h"
+//#include "creport.h"
+//#include "report.h"
 #include "names.h"
-#include "report.h"
 #include "reports.h"
 #include "spells.h"
 #include "wormhole.h"
@@ -66,7 +65,7 @@ void game_init(void)
     register_xmas();
 
 //    register_nr();
-    register_cr();
+//    register_cr();
 
     register_races();
     register_spells();
